Polish & Javadocs
