+++ spring-cloud-aws-autoconfigure/src/main/java/io/awspring/cloud/autoconfigure/config/reload/PollingAwsPropertySourceChangeDetector.java
@@ -16,8 +16,6 @@
 package io.awspring.cloud.autoconfigure.config.reload;
 
 import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
 
 import javax.annotation.PostConstruct;
 
@@ -27,21 +25,28 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
 import org.springframework.core.env.ConfigurableEnvironment;
-import org.springframework.core.env.EnumerablePropertySource;
 import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.support.PeriodicTrigger;
 
-public class PollingSecretsManagerChangeDetector<T extends AwsPropertySource<?, ?>> extends ConfigurationChangeDetector {
+/**
+ * Configuration change detector that checks for changed configuration on a scheduled basis.
+ *
+ * Heavily inspired by Spring Cloud Kubernetes.
+ *
+ * @param <T> - property source class to check
+ * @author Matej Nedic
+ * @author Maciej Walkowiak
+ */
+public class PollingAwsPropertySourceChangeDetector<T extends AwsPropertySource<?, ?>> extends ConfigurationChangeDetector<T> {
 
-	private final Class<T> clazz;
 	protected Log log = LogFactory.getLog(getClass());
 	private final TaskScheduler taskExecutor;
 
-	public PollingSecretsManagerChangeDetector(ReloadableProperties properties, Class<T> clazz, ConfigurationUpdateStrategy strategy,
-			TaskScheduler taskExecutor, ConfigurableEnvironment environment) {
-		super(properties, strategy, environment);
+	public PollingAwsPropertySourceChangeDetector(ReloadableProperties properties, Class<T> clazz,
+		ConfigurationUpdateStrategy strategy,
+		TaskScheduler taskExecutor, ConfigurableEnvironment environment) {
+		super(properties, strategy, environment, clazz);
 		this.taskExecutor = taskExecutor;
-		this.clazz = clazz;
 
 	}
 
@@ -62,7 +67,7 @@ public class PollingSecretsManagerChangeDetector<T extends AwsPropertySource<?, 
 			List<T> currentSecretSources = locateMapPropertySources(this.environment);
 			if (!currentSecretSources.isEmpty()) {
 				for (T propertySource : currentSecretSources) {
-					AwsPropertySource<?,?> clone = propertySource.copy();
+					AwsPropertySource<?, ?> clone = propertySource.copy();
 					clone.init();
 					if (changed(propertySource, clone)) {
 						reloadProperties();
@@ -70,35 +75,5 @@ public class PollingSecretsManagerChangeDetector<T extends AwsPropertySource<?, 
 				}
 			}
 		}
-	}
-
-	/**
-	 * Determines if two property sources are different.
-	 * @param left left map property sources
-	 * @param right right map property sources
-	 * @return {@code true} if source has changed
-	 */
-	public boolean changed(EnumerablePropertySource<?> left, EnumerablePropertySource<?> right) {
-		if (left == right) {
-			return false;
-		}
-		for (String property : left.getPropertyNames()) {
-			if (!Objects.equals(left.getProperty(property), right.getProperty(property))) {
-				return true;
-			}
-		}
-		return false;
-	}
-
-	/**
-	 * Returns a list of MapPropertySource that correspond to the current state of the system. This only handles the
-	 * PropertySource objects that are returned.
-	 * @param environment Spring environment
-	 * @return a list of MapPropertySource that correspond to the current state of the system
-	 */
-	protected List<T> locateMapPropertySources(ConfigurableEnvironment environment) {
-
-		return environment.getPropertySources().stream().filter(it -> (it.getClass().isAssignableFrom(clazz)))
-				.map(it -> (T) it).collect(Collectors.toList());
 	}
 }
