+++ spring-cloud-aws-autoconfigure/src/main/java/io/awspring/cloud/autoconfigure/config/reload/ReloadStrategy.java
@@ -15,6 +15,24 @@
  */
 package io.awspring.cloud.autoconfigure.config.reload;
 
+/**
+ * Configuration reload strategies.
+ *
+ * Heavily inspired by Spring Cloud Kubernetes.
+ *
+ * @author Nicola Ferraro
+ * @author Matej Nedic
+ * @author Maciej Walkowiak
+ */
 public enum ReloadStrategy {
-	REFRESH, RESTART_CONTEXT;
+	/**
+	 * Fire a refresh of beans annotated with @ConfigurationProperties
+	 * or @RefreshScope.
+	 */
+	REFRESH,
+
+	/**
+	 * Restarts the Spring ApplicationContext to apply the new configuration.
+	 */
+	RESTART_CONTEXT
 }
