+++ spring-cloud-aws-autoconfigure/src/main/java/io/awspring/cloud/autoconfigure/config/secretsmanager/SecretsManagerProperties.java
@@ -18,6 +18,7 @@ package io.awspring.cloud.autoconfigure.config.secretsmanager;
 import io.awspring.cloud.autoconfigure.AwsClientProperties;
 import io.awspring.cloud.autoconfigure.config.reload.ReloadProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
 
 /**
  * Configuration properties for the AWS Secrets Manager integration. Mostly based on the Spring Cloud Consul
@@ -38,6 +39,10 @@ public class SecretsManagerProperties extends AwsClientProperties implements Rel
 	 */
 	public static final String CONFIG_PREFIX = "spring.cloud.aws.secretsmanager";
 
+	/**
+	 * Properties related to configuration reload.
+	 */
+	@NestedConfigurationProperty
 	private ReloadProperties reload = new ReloadProperties();
 
 	private boolean monitored;
